Move metadata into its own module
- Moved "tags" to metadata module.
- Moved dataset_description.json generation to a funcdtion within the
  metadata module.
- Added unit tests to assert file is created and content is expected
  (ideally this would be dynamic, but this works for now).
- Updated imports as needed.
