+++ src/niftyone/metadata/__init__.py
@@ -1,0 +1,1 @@
+"""Module for handling metadata associated with dataset or application."""
