+++ tests/unit/niftyone/test_launch.py
@@ -7,7 +7,8 @@ import fiftyone as fo
 import pytest
 from _pytest.logging import LogCaptureFixture
 
-from niftyone import launch, tags
+from niftyone import launch
+from niftyone.metadata import tags
 
 
 @pytest.fixture
@@ -40,7 +41,7 @@ class TestLaunch:
             with (
                 patch("fiftyone.dataset_exists", mock_ds_exists),
                 patch("fiftyone.load_dataset", mock_load_ds),
-                patch("niftyone.tags.GroupTags", mock_group_tags),
+                patch("niftyone.metadata.tags.GroupTags", mock_group_tags),
                 patch("pathlib.Path.exists", mock_path_exists),
                 patch("fiftyone.launch_app", mock_launch_app),
             ):
@@ -72,7 +73,7 @@ class TestLaunch:
             with (
                 patch("fiftyone.dataset_exists", mock_ds_exists),
                 patch("fiftyone.Dataset.from_dir", mock_from_ds_dir),
-                patch("niftyone.tags.GroupTags", mock_group_tags),
+                patch("niftyone.metadata.tags.GroupTags", mock_group_tags),
                 patch("pathlib.Path.exists", mock_path_exists),
                 patch("fiftyone.launch_app", mock_launch_app),
             ):
@@ -102,7 +103,7 @@ class TestLaunch:
         with (
             patch("fiftyone.dataset_exists", mock_ds_exists),
             patch("fiftyone.Dataset.from_dir", mock_from_ds_dir),
-            patch("niftyone.tags.GroupTags", mock_group_tags),
+            patch("niftyone.metadata.tags.GroupTags", mock_group_tags),
             patch("pathlib.Path.exists", mock_path_exists),
             patch("fiftyone.launch_app", mock_launch_app),
         ):
@@ -132,7 +133,7 @@ class TestLaunch:
             with (
                 patch("fiftyone.dataset_exists", mock_ds_exists),
                 patch("fiftyone.load_dataset", mock_load_ds),
-                patch("niftyone.tags.GroupTags", mock_group_tags),
+                patch("niftyone.metadata.tags.GroupTags", mock_group_tags),
                 patch("pathlib.Path.exists", mock_path_exists),
                 patch("fiftyone.launch_app", mock_launch_app),
             ):
@@ -158,7 +159,7 @@ class TestLaunch:
             with (
                 patch("fiftyone.dataset_exists", mock_ds_exists),
                 patch("fiftyone.load_dataset", mock_load_ds),
-                patch("niftyone.tags.GroupTags", mock_group_tags),
+                patch("niftyone.metadata.tags.GroupTags", mock_group_tags),
                 patch("pathlib.Path.exists", mock_path_exists),
                 patch("fiftyone.launch_app", mock_launch_app),
             ):
@@ -188,7 +189,7 @@ class TestLaunch:
             with (
                 patch("fiftyone.dataset_exists", mock_ds_exists),
                 patch("fiftyone.load_dataset", mock_load_ds),
-                patch("niftyone.tags.GroupTags.from_json", mock_group_tags),
+                patch("niftyone.metadata.tags.GroupTags.from_json", mock_group_tags),
                 patch("pathlib.Path.exists", mock_path_exists),
                 patch("fiftyone.launch_app", mock_launch_app),
             ):
